Azure Setup, Configuration and Blob config.
Took a few days, but got the Azure Account set-up and  configured and created the containers to store the blobs (images).

Add Azure dependencies to the main pom and the PhtotoApplication layer's pom. Took a bit of tweeking to get the right format and version, but referencing Microsoft Docs really helped.

Created a class the Class PhotoProperties to get the blob properties from application-blob.properties. All on same layer.

In PhotoServices I created the clients for the Services, container and blob along with connecting them to PhotoProperties.

Also Started with the services to Upload and download the images.
Will do metatdata retrieval next and how to edit that, but for now need get everything working so we can test all the services and move to the UI.