Tag Update, Geolocation on Upload, Miner UI Change
Tag update is now working.

Geolocation was added to upload page so can be added when uploading. Also changed the method that is called from Services.

Home button now takes you to view-album, that way we no get error with back button.
Added location to be displayed with image under tags.