UI Layout
Layouts of all page have been fixed and Final Colour Pallet has been identified and applied.

Only remain problem is the footer and some of the exception handling for some of the pages.

Gonna start with the documentation now and do the final migration in the week when all problems fixed and changes have been made.